+++ fluorender/FluoRender/Formats/msk_reader.cpp
@@ -86,7 +86,7 @@ int MSKReader::LoadBatch(int index)
 
 Nrrd* MSKReader::Convert(int t, int c, bool get_max)
 {
-	size_t pos = m_path_name.find_last_of('.');
+	int64_t pos = m_path_name.find_last_of('.');
 	if (pos == -1)
 		return 0;
 	wstring str_name = m_path_name.substr(0, pos);
