+++ fluorender/FluoRender/Formats/msk_writer.cpp
@@ -65,7 +65,7 @@ void MSKWriter::Save(wstring filename, int mode)
 	if (!m_data)
 		return;
 
-	size_t pos = filename.find_last_of('.');
+	int64_t pos = filename.find_last_of('.');
 	if (pos == -1)
 		return;
 	wstring str_name = filename.substr(0, pos);
